+++ pkg/telnet/export_test.go
@@ -1,0 +1,3 @@
+package telnet
+
+var NewConn = newConn